+++ tests/testthat/test_decode.R
@@ -49,3 +49,14 @@ test_that(
     expect_equal(dt[["word"]][1:8], c("Ja", ",", "ich", "nehme", "die", "Wahl", "an", "."))
   }
 )
+
+test_that(
+  "decode token ids",
+  {
+    gparl <- corpus("GERMAPARLMINI")
+    ids <- RcppCWB::cl_cpos2id(corpus = gparl@corpus, p_attribute = "word", cpos = 0L:(gparl@size - 1L))
+    a <- decode(ids, corpus = gparl, p_attribute = "word", boost = FALSE)
+    b <- decode(ids, corpus = gparl, p_attribute = "word", boost = TRUE)
+    expect_identical(a,b)
+  }
+)